:gear: Déplacer les fonctions getObjectives et collectMissingVariables
