+++ test/rules.test.js
@@ -1,6 +1,6 @@
 import R from 'ramda'
 import {expect} from 'chai'
-import {rules, enrichRule, collectMissingVariables, findVariantsAndRecords2, findVariantsAndRecords, getObjectives} from '../source/engine/rules'
+import {rules, enrichRule, findVariantsAndRecords} from '../source/engine/rules'
 import {analyseSituation} from '../source/engine/traverse'
 
 let stateSelector = (state, name) => null
@@ -22,51 +22,6 @@ describe('enrichRule', function() {
     let rule = {"sous-question":"**wut**"}
     expect(enrichRule(rule)).to.have.property('subquestion','<p><strong>wut</strong></p>\n')
   });
-});
-
-describe('collectMissingVariables', function() {
-
-  it('should derive objectives from the root rule', function() {
-    let rawRules = [
-          {nom: "startHere", formule: {somme: [2, "deux"]}, espace: "sum"},
-          {nom: "deux", formule: 2, "non applicable si" : "sum . evt . ko", espace: "sum"},
-          {nom: "evt", espace: "sum", formule: {"une possibilité":["ko"]}, titre: "Truc", question:"?"},
-          {nom: "ko", espace: "sum . evt"}],
-        rules = rawRules.map(enrichRule),
-        situation = analyseSituation(rules,"startHere")(stateSelector),
-        result = getObjectives(situation)
-
-    expect(result).to.have.lengthOf(1)
-    expect(result[0]).to.have.property('name','deux')
-  });
-
-  it('should identify missing variables', function() {
-    let rawRules = [
-          {nom: "startHere", formule: {somme: [2, "deux"]}, espace: "sum"},
-          {nom: "deux", formule: 2, "non applicable si" : "sum . evt . ko", espace: "sum"},
-          {nom: "evt", espace: "sum", formule: {"une possibilité":["ko"]}, titre: "Truc", question:"?"},
-          {nom: "ko", espace: "sum . evt"}],
-        rules = rawRules.map(enrichRule),
-        situation = analyseSituation(rules,"startHere")(stateSelector),
-        result = collectMissingVariables()(situation)
-
-    expect(result).to.have.property('sum . evt . ko')
-  });
-
-  it('should identify missing variables mentioned in expressions', function() {
-    let rawRules = [
-          {nom: "startHere", formule: {somme: [2, "deux"]}, espace: "sum"},
-          {nom: "deux", formule: 2, "non applicable si" : "evt . nyet > evt . nope", espace: "sum"},
-          {nom: "nope", espace: "sum . evt"},
-          {nom: "nyet", espace: "sum . evt"}],
-        rules = rawRules.map(enrichRule),
-        situation = analyseSituation(rules,"startHere")(stateSelector),
-        result = collectMissingVariables()(situation)
-
-    expect(result).to.have.property('sum . evt . nyet')
-    expect(result).to.have.property('sum . evt . nope')
-  });
-
 });
 
 describe('findVariantsAndRecords', function() {
